+++ philo/srcs/actions_utils.c
@@ -6,13 +6,13 @@
 /*   By: pdubois <pdubois@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/11 01:37:53 by pdubois           #+#    #+#             */
-/*   Updated: 2022/08/11 05:12:47 by pdubois          ###   ########.fr       */
+/*   Updated: 2022/08/12 18:30:10 by pdubois          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-int	ft_take_forks_last_philo(t_info *bag, int name, int *last_meal)
+int	ft_take_forks_last_philo(t_info *bag, int name, long int *last_meal)
 {
 	(void)last_meal;
 	if (pthread_mutex_lock(&(bag->forks[name - 1])) != 0)
